backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
# https://medium.com/django-rest/django-rest-framework-b3028b3f0b9
# https://jkaylight.medium.com/django-rest-framework-authentication-with-dj-rest-auth-4d5e606cde4d
asgiref==3.3.1
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
django-cors-headers==3.7.0
    # via -r requirements.in
django-rest-registration==0.6.0
    # via -r requirements.in
django-rest-swagger==2.2.0
    # via -r requirements.in
django==3.1.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-rest-registration
    #   djangorestframework
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   django-rest-registration
    #   django-rest-swagger
idna==2.10
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
openapi-codec==1.3.2
    # via django-rest-swagger
pytz==2021.1
    # via django
requests==2.25.1
    # via coreapi
simplejson==3.17.2
    # via django-rest-swagger
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.4
    # via requests
django-role-permissions == 3.1.1
djangorestframework-simplejwt ==  4.6.0
drf_yasg == 1.20.0
django-import-export == 2.5.0
django-admin-honeypot == 1.1.0
django_good_otp == 1.2.2
django-admin-env-notice == 0.2
gunicorn == 20.0.4
django-anymail == 8.2
django-maintenance-mode == 0.16.0
python-digitalocean == 1.16.0